@@ -1,35 +1,42 @@
 import React, { useState, useEffect } from 'react';
-import { Route, Link, BrowserRouter } from 'react-router-dom';
-
+import {BrowserRouter,Route,Switch} from 'react-router-dom';
+import NavBar from './NavBar';
+import Component404 from './404';
+import Users from './Users';
+import Home from './Home';
+import Account from './Account';
 import {
-  Products,
   NewProduct,
 } from './index';
-
 import { callApi } from '../util';
+import Products from './Product';
+
 const { REACT_APP_BASE_URL } = process.env;
 
-
 const App = () => {
-    const [token, setToken] = useState('');
-    const [user, setUser] = useState('');
+
     const [products, setProducts] = useState([]);
     const [productId, setProductId] = useState('');
     const [productName, setProductName] = useState([]);
-    const [userId, setUserId] = useState('');
     const [description, setDescription] = useState('');
     const [price, setPrice] = useState('');
-    
+    const [message, setMessage] = useState('');
+    const [token,setToken]=useState(localStorage.getItem("token")||"");
+    const [currentUser,setCurrentUser]=useState(
+    {
+        id:Number(localStorage.getItem("id")),
+        name:localStorage.getItem("username")
+    }||{});
+
     const fetchProducts = async () => {
         try {
             const respObj = await callApi({
-            url: `/products`,
+            url: `products/`,
             token
             });
-            if(respObj)
+            if(respObj&&respObj.allProducts)
             {
-                const productResponse = respObj.data.products;
-                if (productResponse) setProducts(productResponse); 
+                setProducts(respObj.allProducts); 
             }
         } catch (error) {
             throw error;
@@ -46,24 +53,55 @@ const App = () => {
         }
     }, [token]);
 
-    return <div>
-        <header className="site-banner">
-            <BrowserRouter>
-            <Link to='/' className='emblem'><h1>Underground Cars</h1></Link>
-            <div className='nav-bar'>
-                <Link to="/" className="nav-link">Home</Link>
-                <Link to="/products" className="nav-link">Products</Link>
-             
-                {
-                    token ? <Link to='/user/login' className='nav-link' onClick={() => setToken('')}>Log Out</Link> : <Link to='/users/login' className='nav-link'>Login</Link>
-                }
-            </div>
-            </BrowserRouter>
-        </header>
-        <footer />
-    </div>
-}
 
+    return <div className="App">
+
+        <BrowserRouter>
+            <header className="site-banner">
+                <NavBar token={token}></NavBar>
+            </header>
+        
+            <Users setToken={setToken} setCurrentUser={setCurrentUser} currentUser={currentUser}/>
+            
+            <Switch>
+                <Route exact path ="/">
+                    <Home currentUser={currentUser}></Home>
+                </Route>
+
+                <Route exact path ="/products/">
+                    <Products products={products}></Products>
+                </Route>
+
+                {/*
+                Example routes
+
+                <Route exact path ="/">
+                    <Home currentUser={currentUser}></Home>
+                </Route>
+
+                <Route exact path ="/routines/">
+                    <Routines token={token} currentUser={currentUser}></Routines>
+                </Route>
+
+                <Route exact path ="/myroutines/">
+                    <MyRoutines token={token} currentUser={currentUser}></MyRoutines>
+                </Route>
+
+                <Route exact path ="/activities/">
+                    <Activities token={token}></Activities>
+                </Route>
+                */}
+                <Route exact path="/account/">
+                    <Account token={token}></Account>
+                </Route>
+                <Route path="/*">
+                    <Component404></Component404>
+                </Route>
+            </Switch>
+        </BrowserRouter>
+        <footer />
+    </div>;
+}
 export default App;
 
 // import React, { useState, useEffect } from 'react';
@@ -94,4 +132,3 @@ export default App;
 // }
 
 // export default App;
-
