@@ -1,6 +1,6 @@
 require('dotenv').config();
 
-const { REACT_APP_BASE_URL = 'postgres://localhost:5000/' }=process.env.REACT_APP_BASE_URL;
+const { REACT_APP_BASE_URL = 'http://localhost:5000/api/' }=process.env.REACT_APP_BASE_URL;
 
 export const callApi = async ({ url, method, token, body }) => {
     try {
